Борги, раунденд звуки, рантаймы, отключение света от уничтожения АПЦ, БС-аномалии из-за огромного количества блюспейс-майнеров.

+++ code/modules/plumbing/plumbers/autohydro.dm
@@ -14,6 +14,8 @@
 	START_PROCESSING(SSobj, src)
 
 /obj/machinery/hydroponics/constructable/automagic/process()
+	if(!myseed)
+		return
 	if(dead)
 		dead = 0
 		qdel(myseed)
